+++ mp2i/database/models/member.py
@@ -1,6 +1,6 @@
 from typing import Optional
 
-from sqlalchemy import ForeignKey, Integer, PrimaryKeyConstraint
+from sqlalchemy import BigInteger, ForeignKey, Integer, PrimaryKeyConstraint
 from sqlalchemy.orm import Mapped, mapped_column
 
 from . import Base
@@ -15,14 +15,14 @@ class Member(Base):
     __table_args__ = (PrimaryKeyConstraint("guild_id", "user_id"),)
 
     guild_id: Mapped[int] = mapped_column(
-        Integer(),
+        BigInteger(),
         ForeignKey("guilds.guild_id", name="fk_member_guild_id", ondelete="CASCADE"),
         nullable=False,
         info=dict(label="Guild ID", hint="Unique discord ID of the guild."),
     )
 
     user_id: Mapped[int] = mapped_column(
-        Integer(),
+        BigInteger(),
         ForeignKey("users.user_id", name="fk_member_user_id", ondelete="CASCADE"),
         nullable=False,
         info=dict(label="User ID", hint="Unique discord ID of the user."),
